requirements.txt
amqp==2.3.2
asgiref==3.3.1
astroid==2.5.1
Babel==2.6.0
beautifulsoup4==4.9.3
billiard==3.5.0.4
bleach==3.3.0
bootstrap-admin==0.4.0
build==0.3.1.post1
cachetools==2.1.0
celery==4.2.1
certifi==2018.10.15
chardet==3.0.4
check-manifest==0.46
click==7.1.2
cssutils==2.2.0
dateparser==0.7.0
defusedxml==0.5.0
detect-secrets==1.0.3
Django==1.11.29
django-admin-timestamps==1.2
django-appconf==1.0.2
django-autocomplete-light==2.3.6
django-bootstrap-form==3.4
django-bootstrap3==11.0.0
django-celery-beat==1.2.0
django-ckeditor==5.6.1
django-compat==1.0.15
django-compressor==2.2
django-constance==2.3.1
django-custom-field==2.9
django-daterange-filter==1.3.0
django-debug-toolbar==1.10.1
django-environ==0.4.5
django-export-action==0.1.1
django-extensions==2.1.3
django-favicon-plus==0.0.8
django-impersonate==1.3
django-inlinecss==0.3.0
django-js-asset==1.1.0
django-jsonify==0.3.0
django-libsass==0.7
django-localflavor==2.1
django-mass-edit==3.2.0
django-modeltranslation==0.12.2
django-navutils==0.6
django-nose==1.4.7
django-phonenumber-field==2.0.1
django-picklefield==1.1.0
django-positions==0.6.0
django-report-builder==6.4.2
django-responsive-dashboard==1.3.3
django-reversion==3.0.1
django-settings-export==1.2.1
django-simple-import==2.0.2
django-smuggler==0.9.0
django-timezone-field==3.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.4
docutils==0.16
dparse==0.5.1
drf-url-filters==0.5.1
ecdsa==0.16.1
et-xmlfile==1.0.1
flake8==3.9.0
future==0.18.2
gdata-python3==3.0.1
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
httpagentparser==1.8.1
httplib2==0.19.0
idna==2.7
importlib-metadata==3.7.3
isort==5.8.0
jdcal==1.4
kombu==4.2.1
lazy-object-proxy==1.6.0
libsass==0.20.1
lxml==4.6.3
mccabe==0.6.1
nose==1.3.7
oauth2client==4.1.3
oauthlib==2.1.0
openpyxl==2.3.5
packaging==20.9
pdfkit==0.6.1
pep517==0.10.0
persisting-theory==0.2.1
phonenumbers==8.9.16
Pillow==8.1.2
psycopg2-binary==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.7.0
pyflakes==2.3.0
PyJWT==1.6.4
pylint==2.7.2
pynliner==0.8.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-crontab==2.3.5
python-dateutil==2.7.4
python3-openid==3.1.0
pytz==2018.6
PyYAML==5.4.1
raven==6.9.0
rcssmin==1.0.6
redis==2.10.6
regex==2018.8.29
reportlab==3.5.66
requests==2.25.1
requests-oauthlib==1.0.0
rjsmin==1.0.12
rsa==4.7.2
safety==1.10.3
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
soupsieve==2.2.1
sqlparse==0.2.4
tlslite-ng==0.7.6
toml==0.10.2
twilio==6.50.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==1.5.1
Unidecode==1.2.0
uritemplate==3.0.0
urllib3==1.26.4
uWSGI==2.0.17.1
vine==1.1.4
voluptuous==0.11.5
webencodings==0.5.1
wrapt==1.12.1
zipp==3.4.1
